+++ routes/teste.js
@@ -1,12 +1,13 @@
-var express = require('express');
-var router = express.Router();
+const express = require('express');
+const router = express.Router();
+const isLoggedIn = require('../config/isLogged');
 
 /* Página Inicial */
-router.get('/', function(req, res, next) {
-  res.render('teste', { title: 'Teste da Linguinha' });
+router.get('/', isLoggedIn, function(req, res, next) {
+  res.render('modelo', { title: 'Teste da Linguinha', user: req.user });
 });
 
-router.get('/teste', function(req, res, next) {
+router.get('/teste', isLoggedIn, function(req, res, next) {
   res.send("Teste");
 });
