+++ testeDaLinguinha.sublime-workspace
@@ -344,7 +344,7 @@
 		"/home/alcieres/webII/prova2/prova2/app.js",
 		"/home/alcieres/Área de Trabalho/Trabalho/app.js",
 		"/home/alcieres/Área de Trabalho/JS/LOGIN/views/faunamar.ejs",
-		"/home/alcieres/Área de Trabalho/JS/LOGIN/user.js",
+		"/home/alcieres/Área de Trabalho/JS/LOGIN/AssessmentController.js",
 		"/home/alcieres/Área de Trabalho/JS/LOGIN/login.js",
 		"/home/alcieres/Área de Trabalho/Trabalho/views/README.md",
 		"/home/alcieres/Área de Trabalho/Trabalho/views/index.ejs",
