+++ views/assessment/assessmentExtract.ejs
@@ -29,9 +29,6 @@
   </head>
   <body>
 
-  <!--<%- include('../partials/nav', {active: "navAssessment"}); %>-->
-
-
   <div class="container mainContainer"> <%# Containner Relatório Final 1 %>
     <div class="border border-dark mx-auto mb-3">
       <div class="row align-items-center">  <%# Linha do título do Relatório Final %>
