+++ views/modelo.ejs
@@ -19,7 +19,7 @@
   </head>
   <body>
 
-  <%- include('partials/nav', {active: "navAssessment"}); %>
+  <%- include('partials/nav', {active: "home"}); %>
 
     <div class="container mainContainer">
     <p>
